feat: use higlass-schema
