+++ hg/__init__.py
@@ -1,3 +1,4 @@
+from higlass_schema import *
+
 from .api import *
-from .core import *
 from .server import HgServer, server
